Moving to ES2015
