+++ .jshintrc
@@ -10,6 +10,7 @@
     "undef": true,
     "unused": true,
     "browser": true,
+    "esnext": true,
     "globals": {
         "jQuery": true
     }
