+++ package.json
@@ -21,7 +21,10 @@
     "jquery": ">=1.4.3"
   },
   "devDependencies": {
+    "babel-preset-es2015": "^6.3.13",
+    "babelify": "^7.2.0",
     "grunt": "^0.4.5",
+    "grunt-browserify": "^4.0.1",
     "grunt-contrib-clean": "^0.7.0",
     "grunt-contrib-compass": "^1.0.4",
     "grunt-contrib-connect": "^0.11.2",
