+++ src/js/helper.js
@@ -1,0 +1,41 @@
+var helper = {
+    // Check for valids urls
+    // From : http://stackoverflow.com/questions/5717093/check-if-a-javascript-string-is-an-url
+    isUrl(str) {
+      var pattern = new RegExp('^(https?:\\/\\/)?'+ // protocol
+        '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|'+ // domain name
+        '((\\d{1,3}\\.){3}\\d{1,3}))'+ // OR ip (v4) address
+        '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*'+ // port and path
+        '(\\?[;&a-z\\d%_.~+=-]*)?'+ // query string
+        '(\\#[-a-z\\d_]*)?$','i'); // fragment locator
+
+      if (!pattern.test(str)) {
+        return false;
+      } else {
+        return true;
+      }
+    },
+
+    // Loads the content into the menu bar
+    loadContent($menu, content) {
+      $menu.html(content);
+    },
+
+    // Add sidr prefixes
+    addPrefix($element) {
+      var elementId = $element.attr('id'),
+          elementClass = $element.attr('class');
+
+      if (typeof elementId === 'string' && '' !== elementId) {
+        $element.attr('id', elementId.replace(/([A-Za-z0-9_.\-]+)/g, 'sidr-id-$1'));
+      }
+
+      if (typeof elementClass === 'string' && '' !== elementClass && 'sidr-inner' !== elementClass) {
+        $element.attr('class', elementClass.replace(/([A-Za-z0-9_.\-]+)/g, 'sidr-class-$1'));
+      }
+
+      $element.removeAttr('style');
+    }
+};
+
+export default helper;
