+++ src/js/status.js
@@ -1,0 +1,6 @@
+var sidrStatus = {
+    moving: false,
+    opened: false
+};
+
+export default sidrStatus;
